Immutable solution with mutable iterators
The iterator itself is mutable (therefore consumable) but the data
itself is not.

This deliberately copies the original flat data in case I need to use it
again (in the context of a parser). This may not be the case, in which
it can be mutably consumed into a tree.